# Lab9/backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/course_programs": {
            "get": {
                "tags": [
                    "CoursePrograms"
                ],
                "summary": "Get list of course programs",
                "description": "Returns list of course programs",
                "operationId": "601e0233231d5fb30bf61db78a03b695",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CourseProgram"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CoursePrograms"
                ],
                "summary": "Create a new course program",
                "operationId": "storeCourseProgram",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "description": "ID of the course",
                                        "type": "string"
                                    },
                                    "program_id": {
                                        "description": "ID of the program",
                                        "type": "integer"
                                    },
                                    "course_code": {
                                        "description": "Code of the course",
                                        "type": "string"
                                    },
                                    "course_type_id": {
                                        "description": "Type ID of the course",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course program created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseProgram"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/course_programs/{course_program}": {
            "get": {
                "tags": [
                    "CoursePrograms"
                ],
                "summary": "Get a course program by ID",
                "description": "Returns a single course program",
                "operationId": "getCourseProgramById",
                "parameters": [
                    {
                        "name": "course_program",
                        "in": "path",
                        "description": "ID of the course program to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseProgram"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "CoursePrograms"
                ],
                "summary": "Update a course program by ID",
                "description": "Updates the details of a specific course program",
                "operationId": "updateCourseProgramById",
                "parameters": [
                    {
                        "name": "course_program",
                        "in": "path",
                        "description": "ID of the course program to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_code": {
                                        "description": "Code of the course",
                                        "type": "string"
                                    },
                                    "course_type_id": {
                                        "description": "Type ID of the course",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseProgram"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CoursePrograms"
                ],
                "summary": "Delete a course program by ID",
                "description": "Deletes a specific course program",
                "operationId": "deleteCourseProgramById",
                "parameters": [
                    {
                        "name": "course_program",
                        "in": "path",
                        "description": "ID of the course program to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course program deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/course_programs/{course_program}/edit": {
            "get": {
                "tags": [
                    "CoursePrograms"
                ],
                "summary": "Edit a course program by ID",
                "description": "Returns the data of a single course program for editing",
                "operationId": "editCourseProgramById",
                "parameters": [
                    {
                        "name": "course_program",
                        "in": "path",
                        "description": "ID of the course program to edit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseProgram"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course program not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course program not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CourseProgram": {
                "title": "CourseProgram",
                "description": "CourseProgram model schema",
                "required": [
                    "course_id",
                    "program_id",
                    "course_code",
                    "course_type_id"
                ],
                "properties": {
                    "course_id": {
                        "description": "ID of the course",
                        "type": "string",
                        "example": "1"
                    },
                    "program_id": {
                        "description": "ID of the program",
                        "type": "string",
                        "example": "1"
                    },
                    "course_code": {
                        "description": "Code of the course",
                        "type": "string",
                        "example": "CS101"
                    },
                    "course_type_id": {
                        "description": "ID of the course type",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    }
}